refactor: rename venues.py to venue.py
