+++ backend/app/api/__init__.py
@@ -1,5 +1,5 @@
 from app.api.auth import router as auth_router
-from app.api.venues import router as venues_router
+from app.api.venue import router as venue_router
 from app.api.record import router as record_router
 
-__all__ = ["auth_router", "venues_router", "record_router"]
+__all__ = ["auth_router", "venue_router", "record_router"]
