+++ backend/app/main.py
@@ -1,12 +1,12 @@
 from fastapi import FastAPI
 
 from app.api.auth import router as auth_router
-from app.api.venues import router as venues_router
+from app.api.venue import router as venue_router
 from app.api.record import router as record_router
 
 app = FastAPI(title="KULTI API")
 app.include_router(auth_router)
-app.include_router(venues_router)
+app.include_router(venue_router)
 app.include_router(record_router)
